add 'A' (archive and mark read) command to inbox-mode
For some email, the only meaningful content is in the subject line.  It is thus
useful to be able to dispatch this email with a single command after glancing
at it.

As 'N' moves the cursor, and 'a' doesn't clear the "unread" flag, it is
currently necessary to press "Nka" to do this on a single thread.

The 'A' command does this simple task with a single, easy-to-remember,
keystroke, but makes it harder to hit accidentally than 'a'.  A multi_ mode is
supplied.

FossilOrigin-Name: 4aede6284a97bf14fbae6bb3c023bed71afbbf96d2345c061bea91d0aa468bd6