Chronic context fix
Chronic should use a context of :past for email date searches as emails
tend to be in the past. This fixes a problem when the wrong year is
guessed. To search for emails in the future you now need to be less
ambiguous (1 apr 2099 instead of 1 apr).

FossilOrigin-Name: 034df5405ce97771aa605897c0a5c8a62b7a2d467a9302cf9f9988782e746832