Bounce Message Hook
Determine the command used to bounce a message based on a Hook instead
of a configuration option.  Instead of writing an external script that
can send the message properly based on the recipient addresses, rely
on a hook that can return the command based on the From header in the
mail being bounced as well as the intended recipients.  This is more
in line with the sup philosophy.

The default is still to strip any -t from the sendmail command of the
default account.

Signed-off-by: Ben Walton <bwalton@artsci.utoronto.ca>

FossilOrigin-Name: bdeecb205b9ce5b54afb3f149d528b4bd1ce03b3ff5e9651ac5becd2879d769a