rewrap getch in select, handle sigwinch manually
This is necessary to operate with many versions of the ruby ncurses library,
all of which block threads when getch is being called. But this means we can't
use getch to determine a sigwich, since we won't see it until an actual key is
pressed. so we handle sigwinch ourselves.

FossilOrigin-Name: 8cb548592fce59549cb3b33afbc135411ed9be07f2b2cc7c291d641c0746ba00