dont index redundant data
Use the Xapian QueryParser ability to map a prefix in the query to multiple
prefixes in the index. This means we don't need to store duplicate names, email
addresses, and subjects. This also adds the stemmed attachment filenames to the
default (non-prefixed) search. We're storing a subset of the data previous
versions did, so we're able to read them but they can't read us.

FossilOrigin-Name: 4ea266747f20b065aade91086fc742430d4e9a4c21f6cf8dffbfa7ccee850a0f