dont check thread-index-mode dirtiness on every keypress
Now that changes are instantly written to the index it's unnecessary to put the
"modified" message in the status line. Removing this cuts the time for a simple
action like moving the cursor by 75% or more, depending on the number of
messages loaded.

FossilOrigin-Name: 3320f38f253490d232148f9314108b0ff235cd5070f109bcf884659eeb8fb35a