show (recipients) instead of lone "me"
I've changed thread-index-mode to show recipients in () parens instead
of "me" in the From field for threads from me with no replies.

I've developed the habit of archiving a new thread as soon as I send it,
so until I receive a reply it's effectively non-existent.  When checking
on emails I may want to follow up on with a search like "is:sent AND
after:(14 days ago) AND before:(3 days ago)" I would get a pile of lone
"me"'s in the From field.  I would then open each of them to decide
whether or not to follow up based on who it was sent to.

So this helps me cut down on the j<CR>x loops in that scenario.

I'm not sure if this is something that is universally desirable, but
here's the patch anyway.

FossilOrigin-Name: 6f63dbc7291cb95a5517b7546fcd2ce1f504a68e0d146322e91c2324f6a216c2