Use multiple body arrays when calling before-edit for each reply type
This allows for before-edit hooks which modify the body of the message
based on the different headers (previously, they could only modify
headers).

As an example, I use it to sign mail in english if one of the recipients
(not *the* recipient, but one of them) is speaking english.

FossilOrigin-Name: 32fa0dabe4356268134674ad35a1aafa16626675a605c9d9e9f843664a24b5e5