Fixed inline gpg crash when the end marker is missing
lines.between() does not check if the end marker actually exists, but
later it is assumed the marker exists. This change introduces a check
for the end marker and doesn't decrypt messages in which the end marker
is missing. This is a fix for
http://rubyforge.org/pipermail/sup-talk/2010-August/004209.html

FossilOrigin-Name: c84c9c150d4bac704c745edc8c0147f8e5eb0b061cc7f003afda8a651304881a