prefer To and Cc adresses of accounts on reply
Resending this on Rich's request.

If a message was sent (To or Cc header) to an address which corresponds
to an account prefer this account over the account corresponding to the
address in recipient_email.

This solves the problem that the wrong reply from address is chosen on
mails which are sent to an account which is forwarded to another account.

FossilOrigin-Name: 848603f0844ce36f2a9b4d0470020a3ea168b08a798693167c44a9989d93cc9c