fix handling of multiple label: terms in search
By default Xapian will join query terms with the same prefix with OR instead
of AND, so searching for multiple labels doesn't return the expected results.
By making use of a parameter to add_boolean_prefix (added in Xapian 1.2) we
can tell Xapian to use OR only for the search terms that are guaranteed to be
unique.

Signed-off-by: Sascha Silbe <sascha-pgp@silbe.org>

FossilOrigin-Name: 6891ad2d441811dbb17dc79581065cae4de272c7481fbaeaf4b548db296d5e60