Bugfix: Don’t call handle_added_message for old messages with a new location
When sup finds a message, it always called handle_added_message which in turn
invokes add_or_unhide. This is bad, because for killed messages (or marked as
spam), this would bring the message up in the inbox again. After this commit,
handle_added_message is only called when the message is either new or an
existing location of it was updated (to update the labels).

The normal usecase for receiving a message with the same ID is being subscribed
to a mailing list via two different mail addresses.

FossilOrigin-Name: 7b49be45952443b0ab0f27bde7f32dd989050726bb38908f52d092be20a21330