don't leak fds for mbox sources
Only open the mbox when it's actually used and close it again once we're
finished iterating over it.

Adds a method go_idle to Source to inform the source we're unlikely to
use it in the near future so it can do internal clean-up as appropriate.

Signed-off-by: Sascha Silbe <sascha-pgp@silbe.org>

FossilOrigin-Name: e00ec44e94fec42541045e8f37e36d28cf4c6ceeb5ca992787fe509b062bfe0f