Better gpg error checking.
The crypto initialize code now checks for a number of common reasons
for GPG not working, all related to the gpg-agent side of things. It
will then give a reasonably friendly error message, helping diagnose
why GPG is not working for you.

In particular it will check whether the environment variable is set,
whether the environment variable points to a file, and whether that
file is a socket.

FossilOrigin-Name: 80f3871f5d0d1b68934a23342f28fe80730f42587db57a8bf0d08441e7431076