Better gpg error checking.
The crypto initialize code now checks for a number of common reasons
for GPG not working, all related to the gpg-agent side of things. It
will then give a reasonably friendly error message, helping diagnose
why GPG is not working for you.

In particular it will check whether the environment variable is set,
whether the environment variable points to a file, and whether that
file is a socket.

FossilOrigin-Name: 122ecc7b9d8a71eff71039adb45050afc625bf28c5c22740b2616036e560c78e