Shorten confirm discard messages
The messages to confirm discard no longer include the name of the
file being edited. It is still visible on screen, so it is a bit
redundant, and with the file name, the question is too long to fit
on an 80 character terminal.

FossilOrigin-Name: 939032b3d8751d544e952d127f2889cda56dcea646960113019ccbb528629e16