edit-message-mode: Fix the way signatures are handled
The current signatures handling is not suitable for account changing
in edit-message-mode. It is working fine when the edit_signature
option is false, but it could have a better behavior when this option
is enabled.

This commit tries to do this by appending the signature to the body
text if the edit_signature is true, then after editing the message it
checks if the signature has been modified by comparing the end of the
file to the current account's signature. If it has been edited then we
stick to it by setting @sig_edited to true, otherwise the signature
will still be automatically changed if another account is selected.

Note the code was originally written by Damien Leone
<damien.leone@fensalir.fr> but the patch did not apply when I got round
to doing it, and needed some refactoring for the current state of the
code.

FossilOrigin-Name: dc6bae77caa44f146174396e1b17f57340ada52a05837a29096dbda38a15f1eb