reply-mode: improve the way headers are handled
I noticed that changing the Reply-to selector was also changing the
other headers such as From even if you edited it. I guess it is an
unwanted behaviour since this selector should only behave on the To
and Cc fields.

This commit splits the headers in two "types", those handled by
Reply-To selector and those that are only merged with the first ones
at initialization of the edit-mode.

It also gives a better support of the Customized choice by saving back
the fields for later use.

Signed-off-by: Hamish Downer <dmishd@gmail.com>

FossilOrigin-Name: 92335c839804b25f96572d898f097cd445730e31d402f9abb10b51233421fe4c