Synchronize remote modifications from a Maildir source to sup
It is now possible to use multiple clients to handle Maildirs (and
IMAP server if you use offlineimap).

It works by using files' ctime instead of mtime when polling so we can
detect if the mail's flags have changed. Maildir#poll now appends
"subdir/ids" in the added and deleted arrays at each loop instead of
recreating them. Arrays are then compared to each other allowing us to
see if a mail has been updated or moved (from "cur/" to "new/" for
instance, this is what offlineimap does when you mark a read mail as
unread on an IMAP server).

The index is then updated and the display is refreshed. When a mail is
deleted it is now *really* deleted from xapian. Before, only its
location was removed but the mail could still be visible in search
results although we could'nt load its content ("An error occurred
while loading this message").

Beware that sup does NOT synchronize its modifications to the source.

FossilOrigin-Name: 0c565af2ead63e90c2eae502f577ee8779a58b119677b15ce56633146cf94640